Splitting out a new class `NamespaceNodeToReflectionTypeContext`
